refactor: Update FirebaseApp class to use authentication context
The FirebaseApp class has been updated to use the authentication context provided by the AuthProvider. The 'auth' property has been added to the class and is now initialized in the constructor using the getAuth() function. This allows for easier access to the authentication methods throughout the class.

This commit refactors the code to improve the organization and readability of the FirebaseApp class, and prepares it for further implementation of the authentication functionality.