refactor: Update FirebaseApp class to use signInWithRedirect for Google sign-in
The FirebaseApp class has been updated to use the signInWithRedirect method for Google sign-in. This change improves the authentication flow by redirecting the user to the Google sign-in page instead of using a popup. The getRedirectResult method is now called after the sign-in process to retrieve the authentication response. This refactor prepares the code for further implementation of the authentication functionality.