refactor: Update login method to support multiple providers
The `login` method in the `App` component has been updated to support multiple authentication providers. Users can now sign in with either Google or Facebook accounts by passing the corresponding provider as an argument to the `login` function. This change improves the authentication flow and provides users with more options for signing in.